+++ cmd/fablab/main.go
@@ -22,9 +22,9 @@ import (
 	"github.com/netfoundry/fablab/kernel/model"
 	"github.com/netfoundry/fablab/zitilib"
 	_ "github.com/netfoundry/fablab/zitilib"
-	_ "github.com/netfoundry/fablab/zitilib/characterization"
-	_ "github.com/netfoundry/fablab/zitilib/examples"
-	_ "github.com/netfoundry/fablab/zitilib/mattermozt"
+	_ "github.com/netfoundry/fablab/zitilib/models/characterization"
+	_ "github.com/netfoundry/fablab/zitilib/models/examples"
+	_ "github.com/netfoundry/fablab/zitilib/models/mattermozt"
 	"github.com/netfoundry/ziti-foundation/transport"
 	"github.com/netfoundry/ziti-foundation/transport/quic"
 	"github.com/netfoundry/ziti-foundation/transport/tcp"
