+++ zitilib/models/characterization/actions/stop.go
