+++ zitilib/models/characterization/activation.go
