+++ zitilib/models/characterization/infrastructure.go
