+++ zitilib/models/characterization/model.go
