+++ zitilib/models/examples/configuration.go
@@ -14,22 +14,30 @@
 	limitations under the License.
 */
 
-package mattermozt
+package zitilib_examples
 
 import (
 	"github.com/netfoundry/fablab/kernel/fablib/runlevel/1_configuration/config"
 	"github.com/netfoundry/fablab/kernel/model"
-	"github.com/netfoundry/fablab/zitilib/runlevel/1_configuration/pki"
+	"github.com/netfoundry/fablab/zitilib/runlevel/1_configuration"
 )
 
 func newConfigurationFactory() model.Factory {
 	return &configurationFactory{}
 }
 
-func (self *configurationFactory) Build(m *model.Model) error {
+func (_ *configurationFactory) Build(m *model.Model) error {
 	m.Configuration = model.ConfigurationBinders{
-		func(m *model.Model) model.ConfigurationStage { return pki.Group(pki.Fabric(), pki.DotZiti()) },
+		func(m *model.Model) model.ConfigurationStage { return zitilib_runlevel_1_configuration.IfNoPki(zitilib_runlevel_1_configuration.Fabric(), zitilib_runlevel_1_configuration.DotZiti()) },
 		func(m *model.Model) model.ConfigurationStage { return config.Component() },
+		func(m *model.Model) model.ConfigurationStage {
+			configs := []config.StaticConfig{
+				{Src: "loop/10-ambient.loop2.yml", Name: "10-ambient.loop2.yml"},
+				{Src: "loop/4k-chatter.loop2.yml", Name: "4k-chatter.loop2.yml"},
+				{Src: "remote_identities.yml", Name: "remote_identities.yml"},
+			}
+			return config.Static(configs)
+		},
 	}
 	return nil
 }
