+++ zitilib/models/mattermozt/configuration.go
@@ -14,21 +14,21 @@
 	limitations under the License.
 */
 
-package zitilib_characterization
+package mattermozt
 
 import (
 	"github.com/netfoundry/fablab/kernel/fablib/runlevel/1_configuration/config"
 	"github.com/netfoundry/fablab/kernel/model"
-	"github.com/netfoundry/fablab/zitilib/runlevel/1_configuration/pki"
+	"github.com/netfoundry/fablab/zitilib/runlevel/1_configuration"
 )
 
 func newConfigurationFactory() model.Factory {
 	return &configurationFactory{}
 }
 
-func (f *configurationFactory) Build(m *model.Model) error {
+func (self *configurationFactory) Build(m *model.Model) error {
 	m.Configuration = model.ConfigurationBinders{
-		func(m *model.Model) model.ConfigurationStage { return pki.Group(pki.Fabric(), pki.DotZiti()) },
+		func(m *model.Model) model.ConfigurationStage { return zitilib_runlevel_1_configuration.IfNoPki(zitilib_runlevel_1_configuration.Fabric(), zitilib_runlevel_1_configuration.DotZiti()) },
 		func(m *model.Model) model.ConfigurationStage { return config.Component() },
 	}
 	return nil
